+++ Cargo.toml
@@ -1,5 +1,5 @@
 [workspace]
-members = ["payjoin", "payjoin-cli", "payjoin-directory", "payjoin-defaults"]
+members = ["payjoin", "payjoin-cli", "payjoin-directory"]
 resolver = "2"
 
 [patch.crates-io.payjoin]
