+++ payjoin-cli/Cargo.toml
@@ -19,8 +19,8 @@ path = "src/main.rs"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 [features]
 native-certs = ["reqwest/rustls-tls-native-roots"]
-danger-local-https = ["rcgen", "reqwest/rustls-tls", "rustls", "hyper-rustls", "payjoin-defaults/danger-local-https"]
-v2 = ["payjoin/v2", "payjoin-defaults/v2"]
+danger-local-https = ["rcgen", "reqwest/rustls-tls", "rustls", "hyper-rustls", "payjoin/danger-local-https"]
+v2 = ["payjoin/v2", "payjoin/io"]
 
 [dependencies]
 anyhow = "1.0.70"
@@ -34,7 +34,6 @@ hyper = { version = "0.14", features = ["full"] }
 hyper-rustls = { version = "0.25", optional = true }
 log = "0.4.7"
 payjoin = { version = "0.15.0", features = ["send", "receive", "base64"] }
-payjoin-defaults = { path = "../payjoin-defaults" }
 rcgen = { version = "0.11.1", optional = true }
 reqwest = { version = "0.12", default-features = false }
 rustls = { version = "0.22.2", optional = true }
