--- payjoin-defaults/Cargo.toml
@@ -1,37 +1,0 @@
-[package]
-authors = ["jbesraa <jbesraa@gmail.com>"]
-description = "IO utilities for Payjoin"
-edition = "2021"
-keywords = ["bip77", "payjoin", "bitcoin", "networking"]
-license = "MITNFA"
-name = "payjoin-defaults"
-readme = "README.md"
-repository = "https://github.com/payjoin/rust-payjoin"
-resolver = "2"
-version = "0.0.1"
-
-[features]
-danger-local-https = ["reqwest/rustls-tls", "rustls"]
-v2 = ["payjoin/v2"]
-
-[dependencies]
-payjoin = { version = "0.15.0", features = ["send", "receive"] }
-rustls = { version = "0.22.2", optional = true }
-reqwest = { version = "0.12", default-features = false }
-
-[dev-dependencies]
-bitcoin = { version = "0.30.0", features = ["base64"] }
-bitcoind = { version = "0.31.1", features = ["0_21_2"] }
-http = "1"
-log = { version = "0.4.14"}
-ohttp-relay = "0.0.8"
-once_cell = "1"
-payjoin-directory = { path = "../payjoin-directory", features = ["danger-local-https"] }
-rcgen = { version = "0.11" }
-rustls = "0.22.2"
-testcontainers = "0.15.0"
-testcontainers-modules = { version = "0.1.3", features = ["redis"] }
-tokio = { version = "1.12.0", features = ["full"] }
-tracing = "0.1.40"
-tracing-subscriber = { version = "0.3.17", features = ["env-filter"] }
-url = "2.2.2"
