--- payjoin-defaults/README.md
@@ -1,9 +1,0 @@
-# payjoin-defaults
-
-This provides a collection of I/O utilities for working with the
-payjoin crate.  The payjoin crate only deals with the low-level
-protocol. payjoin-defaults provides sane defaults to make implementing
-payjoin easy.
-
-To learn more about payjoin, refer to [Payjoin Website](https://payjoin.org/).
-
